pyproject.toml
[project]
name = "promt-eval"
version = "0.1.0"
description = "Injected prompt evaluation tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.8.0",
    "anthropic>=0.8.0",
    "click>=8.0.0",
    "docx2pdf>=0.1.8",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "google-generativeai>=0.3.0",
    "numpy>=1.24.0",
    "openai>=1.3.0",
    "pandas>=2.0.0",
    "pypdf2>=3.0.1",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "sentence-transformers>=2.0.0",
    "sqlalchemy>=2.0.0",
    "tqdm>=4.65.0",
    "werkzeug>=2.3.0",
]
